+++ ui/app/src/__mocks__/server.js
@@ -1,4 +1,4 @@
+import { getChaosMeshDashboardAPIMSW } from '@/openapi/index.msw'
 import { setupServer } from 'msw/node'
-import { getChaosMeshDashboardAPIMSW } from 'openapi/index.msw'
 
 export const server = setupServer(...getChaosMeshDashboardAPIMSW())