+++ ui/app/src/components/Scope/DeprecatedAddress.tsx
@@ -16,8 +16,8 @@
  */
 import { getIn, useFormikContext } from 'formik'
 
-import { LabelField } from 'components/FormField'
-import i18n from 'components/T'
+import { LabelField } from '@/components/FormField'
+import i18n from '@/components/T'
 
 const DeprecatedAddress = () => {
   const { errors, touched } = useFormikContext()