+++ ui/app/src/components/Scope/TargetsTable/PodsTable.tsx
@@ -14,12 +14,11 @@
  * limitations under the License.
  *
  */
+import PaperContainer from '@/mui-extends/PaperContainer'
+import { TypesPod } from '@/openapi/index.schemas'
 import { Checkbox, Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from '@mui/material'
-import { TypesPod } from 'openapi/index.schemas'
 
-import PaperContainer from '@ui/mui-extends/esm/PaperContainer'
-
-import { T } from 'components/T'
+import { T } from '@/components/T'
 
 import { TargetsTableActions } from '.'