+++ ui/app/src/components/TopContainer/index.tsx
@@ -14,7 +14,13 @@
  * limitations under the License.
  *
  */
-import loadable from '@loadable/component'
+import { applyAPIAuthentication, applyErrorHandling, applyNSParam } from '@/api/interceptors'
+import { Stale } from '@/api/queryUtils'
+import ConfirmDialog from '@/mui-extends/ConfirmDialog'
+import Loading from '@/mui-extends/Loading'
+import { useGetCommonConfig } from '@/openapi'
+import { useAuthActions, useAuthStore } from '@/zustand/auth'
+import { useComponentActions, useComponentStore } from '@/zustand/component'
 import {
   Alert,
   Box,
@@ -28,30 +34,20 @@ import {
   useTheme,
 } from '@mui/material'
 import { styled } from '@mui/material/styles'
-import { applyAPIAuthentication, applyNSParam } from 'api/interceptors'
-import { Stale } from 'api/queryUtils'
 import Cookies from 'js-cookie'
-import { useGetCommonConfig } from 'openapi'
-import { useEffect, useState } from 'react'
-import { Outlet } from 'react-router-dom'
+import { lazy, useEffect, useState } from 'react'
+import { Outlet } from 'react-router'
 
-import ConfirmDialog from '@ui/mui-extends/esm/ConfirmDialog'
-import Loading from '@ui/mui-extends/esm/Loading'
+import { TokenFormValues } from '@/components/Token'
 
-import { useStoreDispatch, useStoreSelector } from 'store'
-
-import { setAlertOpen, setAuthOpen, setConfirmOpen, setNameSpace, setTokenName, setTokens } from 'slices/globalStatus'
-
-import { TokenFormValues } from 'components/Token'
-
-import insertCommonStyle from 'lib/d3/insertCommonStyle'
-import LS from 'lib/localStorage'
+import insertCommonStyle from '@/lib/d3/insertCommonStyle'
+import LS from '@/lib/localStorage'
 
 import Navbar from './Navbar'
 import { closedWidth, openedWidth } from './Sidebar'
 import Sidebar from './Sidebar'
 
-const Auth = loadable(() => import('./Auth'))
+const Auth = lazy(() => import('./Auth'))
 
 const Root = styled(Box, {
   shouldForwardProp: (prop) => prop !== 'open',
@@ -72,11 +68,13 @@ const Root = styled(Box, {
 const TopContainer = () => {
   const theme = useTheme()
 
-  const { alert, alertOpen, confirm, confirmOpen, authOpen } = useStoreSelector((state) => state.globalStatus)
-
-  const dispatch = useStoreDispatch()
-  const handleSnackClose = () => dispatch(setAlertOpen(false))
-  const handleConfirmClose = () => dispatch(setConfirmOpen(false))
+  const alert = useComponentStore((state) => state.alert)
+  const alertOpen = useComponentStore((state) => state.alertOpen)
+  const confirm = useComponentStore((state) => state.confirm)
+  const confirmOpen = useComponentStore((state) => state.confirmOpen)
+  const { setAlert, setAlertOpen, setConfirmOpen } = useComponentActions()
+  const authOpen = useAuthStore((state) => state.authOpen)
+  const { setAuthOpen, setNameSpace, setTokenName, setTokens, removeToken } = useAuthActions()
 
   // Sidebar related
   const miniSidebar = LS.get('mini-sidebar') === 'y'
@@ -88,61 +86,64 @@ const TopContainer = () => {
 
   const [loading, setLoading] = useState(true)
 
-  /**
-   * Set authorization (RBAC token / GCP) for API use.
-   *
-   */
-  function setAuth() {
-    // GCP
-    const accessToken = Cookies.get('access_token')
-    const expiry = Cookies.get('expiry')
-
-    if (accessToken && expiry) {
-      const token = {
-        accessToken,
-        expiry,
-      }
-
-      applyAPIAuthentication(token as any)
-      dispatch(setTokenName('gcp'))
-
-      return
-    }
-
-    const token = LS.get('token')
-    const tokenName = LS.get('token-name')
-    const globalNamespace = LS.get('global-namespace')
-
-    if (token && tokenName) {
-      const tokens: TokenFormValues[] = JSON.parse(token)
-
-      applyAPIAuthentication(tokens.find(({ name }) => name === tokenName)!.token)
-      dispatch(setTokens(tokens))
-      dispatch(setTokenName(tokenName))
-    } else {
-      dispatch(setAuthOpen(true))
-    }
-
-    if (globalNamespace) {
-      applyNSParam(globalNamespace)
-      dispatch(setNameSpace(globalNamespace))
-    }
-  }
-
-  useGetCommonConfig({
+  const { data } = useGetCommonConfig({
     query: {
       staleTime: Stale.DAY,
-      onSuccess(data) {
-        if (data.security_mode) {
-          setAuth()
-        }
-
-        setLoading(false)
-      },
     },
   })
 
   useEffect(() => {
+    /**
+     * Set authorization (RBAC token / GCP) for API use.
+     */
+    function setAuth() {
+      // GCP
+      const accessToken = Cookies.get('access_token')
+      const expiry = Cookies.get('expiry')
+
+      if (accessToken && expiry) {
+        const token = {
+          accessToken,
+          expiry,
+        }
+
+        applyAPIAuthentication(token)
+        setTokenName('gcp')
+
+        return
+      }
+
+      const token = LS.get('token')
+      const tokenName = LS.get('token-name')
+      const globalNamespace = LS.get('global-namespace')
+
+      if (token && tokenName) {
+        const tokens: TokenFormValues[] = JSON.parse(token)
+
+        applyAPIAuthentication(tokens.find(({ name }) => name === tokenName)!.token)
+        setTokens(tokens)
+        setTokenName(tokenName)
+      } else {
+        setAuthOpen(true)
+      }
+
+      if (globalNamespace) {
+        applyNSParam(globalNamespace)
+        setNameSpace(globalNamespace)
+      }
+    }
+
+    if (data) {
+      if (data.security_mode) {
+        setAuth()
+      }
+
+      setLoading(false)
+    }
+  }, [data])
+
+  useEffect(() => {
+    applyErrorHandling({ openAlert: setAlert, removeToken })
     insertCommonStyle()
   }, [])
 
@@ -178,9 +179,9 @@ const TopContainer = () => {
           }}
           autoHideDuration={6000}
           open={alertOpen}
-          onClose={handleSnackClose}
+          onClose={() => setAlertOpen(false)}
         >
-          <Alert severity={alert.type} onClose={handleSnackClose}>
+          <Alert severity={alert.type} onClose={() => setAlertOpen(false)}>
             {alert.message}
           </Alert>
         </Snackbar>
@@ -189,7 +190,7 @@ const TopContainer = () => {
       <Portal>
         <ConfirmDialog
           open={confirmOpen}
-          close={handleConfirmClose}
+          close={() => setConfirmOpen(false)}
           title={confirm.title}
           description={confirm.description}
           onConfirm={confirm.handle}